Move providers to module
